Implement transfer of control
This commit updates old documentation, figures out the layout of the
control record in x86-64, implements the control transfer mechanism in
inline assembly, renames the status concept to control block, and adds a
lot of debugging information (which will go away soon).

The final return from the coroutine is currently broken due to the
coroutine function overwriting the contents in the return instruction
pointer. I think we can take advantage of the free padding space to
separate the resume and return pointers into two fields.