Copy the `FnOnce` instance to the new stack.
This commit includes many changes, because I noticed that keeping a
reference to the `FnOnce` instance in the old stack, and copying it from
the trampoline is not safe: even if the `Coro::new` function does not
drop `f`, its caller is free to clobber the stack containing `f` prior
to activating the coroutine. The fix required a major change to the
control structure, and it now includes a new mechanism to detect
situations when `yield` is called from outside a coroutine.

The good news is that the `simple` test now passes! This commit also
adds a control transfer mechanism optimized for the last return of a
coroutine.